pyproject.toml
[tool.poetry]
name = "elevator-dqn"
version = "0.1.0"
description = ""
authors = [""]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.5"
mxnet = "^1.5"
pyyaml = "^5.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
